+++ packages/config/src/index.ts
@@ -18,7 +18,7 @@
 
 export { HOMEPAGE_MILESTONES } from './global/milestones'
 export { PROJECT_COUNTDOWNS } from './global/countdowns'
-export { ProjectService, type Project } from './projects/project/ProjectService'
+export { ProjectService, type Project } from './ProjectService'
 export {
   safeGetTokenByAssetId,
   tokenList,
